CCB-007 Modularize query class for each repository
