+++ src/modules/financial-asset/repository/instrument.repository.ts
@@ -1,9 +1,15 @@
+import { Repository } from 'typeorm';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, SelectQueryBuilder } from 'typeorm';
 
-import { IInstrumentRepository } from '@financial-asset/repository';
+import { InstrumentQuery } from '@financial-asset/repository';
 import { IInstrumentEntity, InstrumentEntity } from '@financial-asset/entity';
+
+export interface IInstrumentRepository {
+  create: (data: IInstrumentEntity) => Promise<IInstrumentEntity>;
+  find: (query: { text: string }) => Promise<IInstrumentEntity[]>;
+  findOne: (query: Partial<IInstrumentEntity>) => Promise<IInstrumentEntity>;
+}
 
 @Injectable()
 export class InstrumentRepository implements IInstrumentRepository {
@@ -31,37 +37,5 @@ export class InstrumentRepository implements IInstrumentRepository {
     query.byTicker(queries.ticker);
 
     return query.getOne();
-  }
-}
-
-export class InstrumentQuery {
-  private queryBuilder: SelectQueryBuilder<InstrumentEntity>;
-
-  constructor(private readonly repository: Repository<InstrumentEntity>) {
-    this.queryBuilder = this.repository.createQueryBuilder('i');
-  }
-
-  byTicker(ticker: string): this {
-    if (!ticker?.length) return this;
-
-    this.queryBuilder.orWhere('i.ticker ILIKE :ticker', { ticker: `%${ticker}%` });
-
-    return this;
-  }
-
-  byName(name: string): this {
-    if (!name?.length) return this;
-
-    this.queryBuilder.orWhere('i.name ILIKE :name', { name: `%${name}%` });
-
-    return this;
-  }
-
-  async getMany(): Promise<IInstrumentEntity[]> {
-    return await this.queryBuilder.getMany();
-  }
-
-  async getOne(): Promise<IInstrumentEntity | undefined> {
-    return await this.queryBuilder.getOne();
   }
 }
