+++ src/modules/market/repository/index.ts
@@ -1,1 +1,2 @@
 export * from './market.repository';
+export * from './market.query';
