+++ src/modules/order/repository/index.ts
@@ -1,1 +1,2 @@
 export * from './order.repository';
+export * from './order.query';
