+++ src/modules/order/repository/order.repository.ts
@@ -1,8 +1,9 @@
+import { Repository } from 'typeorm';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, SelectQueryBuilder } from 'typeorm';
 
-import { EOrderStatuses, IOrderEntity, OrderEntity } from '@order/entity';
+import { OrderQuery } from '@order/repository';
+import { IOrderEntity, OrderEntity } from '@order/entity';
 
 export interface IOrderRepository {
   find: (query: Partial<IOrderEntity>) => Promise<IOrderEntity[]>;
@@ -25,45 +26,5 @@ export class OrderRepository implements IOrderRepository {
 
   async create(data: IOrderEntity): Promise<IOrderEntity> {
     return this.repository.save(this.repository.create(data));
-  }
-}
-
-export class OrderQuery {
-  private queryBuilder: SelectQueryBuilder<OrderEntity>;
-
-  constructor(private readonly repository: Repository<OrderEntity>) {
-    this.queryBuilder = this.repository.createQueryBuilder('o');
-  }
-
-  byUserId(id: number): this {
-    if (!id) return this;
-
-    this.queryBuilder.orWhere('o.userid = :id', { id });
-
-    return this;
-  }
-
-  byStatus(status: EOrderStatuses): this {
-    if (!status) return this;
-
-    this.queryBuilder.andWhere('o.status = :status', { status });
-
-    return this;
-  }
-
-  orderByCreated(type) {
-    this.queryBuilder.orderBy('o.datetime', type);
-
-    return this;
-  }
-
-  async getMany(): Promise<IOrderEntity[]> {
-    this.queryBuilder.leftJoinAndSelect('o.instrumentid', 'instrument');
-
-    return await this.queryBuilder.getMany();
-  }
-
-  async getOne(): Promise<IOrderEntity | undefined> {
-    return await this.queryBuilder.getOne();
   }
 }
