+++ src/modules/user/repository/index.ts
@@ -1,1 +1,2 @@
 export * from './user.repository';
+export * from './user.query';
