+++ src/modules/user/repository/user.repository.ts
@@ -1,7 +1,8 @@
+import { Repository } from 'typeorm';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, SelectQueryBuilder } from 'typeorm';
 
+import { UserQuery } from '@user/repository';
 import { IUserEntity, UserEntity } from '@user/entity';
 
 export interface IUserRepository {
@@ -26,29 +27,5 @@ export class UserRepository implements IUserRepository {
     query.byId(queries.id);
 
     return query.getOne();
-  }
-}
-
-export class UserQuery {
-  private queryBuilder: SelectQueryBuilder<UserEntity>;
-
-  constructor(private readonly repository: Repository<UserEntity>) {
-    this.queryBuilder = this.repository.createQueryBuilder('u');
-  }
-
-  byId(id: string): this {
-    if (!id) return this;
-
-    this.queryBuilder.orWhere('u.id = :id', { id });
-
-    return this;
-  }
-
-  async getMany(): Promise<IUserEntity[]> {
-    return await this.queryBuilder.getMany();
-  }
-
-  async getOne(): Promise<IUserEntity | undefined> {
-    return await this.queryBuilder.getOne();
   }
 }
